refactor(liveslots-tools): extract fake storage tools
